requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.6.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmdstanpy==1.3.0
contourpy==1.3.3
cycler==0.12.1
Django==4.1.6
django-cors-headers==3.14.0
django-redis==5.4.0
djangorestframework==3.14.0
dnspython==2.8.0
fastapi==0.128.0
fonttools==4.61.1
google-auth==2.43.0
google-genai==1.52.0
h11==0.16.0
holidays==0.87
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
joblib==1.5.3
kagglehub==0.4.1
kagglesdk==0.1.15
kiwisolver==1.4.9
matplotlib==3.10.8
mysqlclient==2.1.1
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prophet==1.2.1
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.6.0
pymongo==4.15.5
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.7
python-dotenv==1.2.1
pytz==2022.7.1
PyYAML==6.0.3
razorpay==2.0.0
redis==7.1.0
requests==2.28.2
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.3
stanio==0.5.1
starlette==0.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2022.7
ua-parser==0.18.0
urllib3==1.26.15
user-agents==2.2.0
uvicorn==0.40.0
websockets==15.0.1
xgboost==3.1.2
